Adding links to talk by Javier Maestro and David Pope of Facebook that was missing from 2016 hbasecon operations track at Javier's request... see below:
From: "J. Javier Maestro" <jjmaestro@ieee.org>
Date: Fri, 3 Dec 2021 13:38:12 +0000
Message-ID: <CABvji04C66wr4niSCEEc9gNib_ah9j995nw1EOEnwnpAmhXVuA@mail.gmail.com>
Subject: An email from time past!
To: stack@duboce.net

Hi Michael!

We briefly met at HBaseCon West 2016, I used to work as a Production
Engineer in the HBase Facebook team with Elliot Clark, Mikhail Antonov,
etc. I'm writing out of the blue with a weird request for some help to mend
the HBaseCon pages =F0=9F=98=85

I (together with another co-worker) presented a talk on containerizing
HBase. I don't particularly enjoy presenting in conferences, so I've only
done it two times in my professional career. However, that HBase talk went
quite well and I was happy with it. So, fast-forward a few years, I went to
refresh my LinkedIn and... when I went looking in the HBaseCon website /
archive page for the record of the talk, there was nothing to be found. It
was gone, as if we never gave the talk.

I remember that, at the time, we were told that some talks would not be
recorded because they had some issues / complaints from Facebook in the
past... and that's my best guess as to what happened to the talk. Not being
recorded, I guess someone just removed the entry from the table thinking
that it might have never happened.

I feel a bit silly asking for some help about this, since it's something
pretty small and insignificant, but at the same time, since I've never done
many public talks, it did make me feel a bit sad that the only official
record of it is now gone like tears in the rain XXD

I tried a couple years ago to send an email to dev@hbase.apache.org but I
never heard back. Since you have been so involved with the HBase project
and the community, I thought you might know someone I could talk to
directly and see if they could add the record of the talk to the HBaseCon
and Archives pages (http://hbase.apache.org/www.hbasecon.com/ and
https://hbase.apache.org/hbasecon-archives.html). The slides from the talk
are here:
https://speakerdeck.com/jjmaestro/hbasecon-2016-west-containerizing-apache-
hbase-clusters

Again, sorry for the cold-email and thanks a ton in advance for your time
and help!

Cheers,

Javier